Create Best Time to Buy and Sell Stock II.cpp
You are given an integer array of prices where prices[i] is the price of a given stock on an ith day.

On each day, you may decide to buy and/or sell the stock. You can only hold at most one share of the stock at any time. However, you can buy it and then immediately sell it on the same day.

Find and return the maximum profit you can achieve.

A space and time-optimized solution are proposed.